requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.14
django-allauth==0.43.0
django-countries==7.3.2
django-crispy-forms==1.10.0
django-debug-toolbar==3.2.4
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.3
oauthlib==3.2.0
paystackapi==2.1.0
Pillow==2.2.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyconfig==3.2.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytool==3.16.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.10
whitenoise==6.2.0
